styling improvements
