+++ rental-app-frontend/app/(tabs)/profile.tsx
@@ -1,0 +1,51 @@
+import React from "react";
+import { View, Text, StyleSheet, Image } from "react-native";
+
+export default function ProfileScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.headerTitle}>Profile</Text>
+
+      <View style={styles.profileContent}>
+      <Image
+          source={{ uri: "https://via.placeholder.com/150" }}
+          style={styles.avatar}
+        />
+        <Text style={styles.name}>John Doe</Text>
+        <Text style={styles.email}>johndoe@example.com</Text>
+      </View>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: "#f5f5f5",
+  },
+  headerTitle: {
+    fontSize: 24,
+    fontWeight: "bold",
+    textAlign: "left",
+    marginBottom: 16,
+  },
+  profileContent: {
+    alignItems: "center",
+  },
+  avatar: {
+    width: 150,
+    height: 150,
+    borderRadius: 75,
+  },
+  name: {
+    marginTop: 20,
+    fontSize: 24,
+    fontWeight: "600",
+  },
+  email: {
+    marginTop: 10,
+    fontSize: 18,
+    color: "gray",
+  },
+});
