+++ rental-app-frontend/components/navigation/TabBarIcon.tsx
@@ -1,5 +1,3 @@
-// You can explore the built-in icon families and icons on the web at https://icons.expo.fyi/
-
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { type IconProps } from '@expo/vector-icons/build/createIconSet';
 import { type ComponentProps } from 'react';
